requirements.txt
absl-py==2.1.0
altair==5.5.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
anyio==4.4.0
archspec @ file:///croot/archspec_1709217642129/work
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.7.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout @ file:///croot/async-timeout_1703096998144/work
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.2
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chex==0.1.89
click==8.1.8
comm==0.2.2
conda @ file:///croot/conda_1715635703388/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1714483155348/work
conda_package_streaming @ file:///work/perseverance-python-buildout/croot/conda-package-streaming_1698847176583/work
contourpy==1.2.1
cryptography @ file:///croot/cryptography_1714660666131/work
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated @ file:///work/perseverance-python-buildout/croot/deprecated_1698868068243/work
distro @ file:///croot/distro_1714488253808/work
dm-tree==0.1.9
docstring_parser==0.16
einops==0.8.1
etils==1.12.2
executing==2.0.1
facenet-pytorch==2.6.0
fastjsonschema==2.20.0
filelock==3.18.0
flatbuffers==25.2.10
flax==0.10.4
fonttools==4.53.0
fqdn==1.5.1
frozendict @ file:///croot/frozendict_1713194832637/work
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.0.3
grpcio==1.70.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.12.2
idna @ file:///croot/idna_1714398848350/work
imbalanced-learn==0.13.0
immutabledict==4.2.1
importlib_resources==6.5.2
ipykernel==6.29.4
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
jwcrypto @ file:///home/conda/feedstock_root/build_artifacts/jwcrypto_1709770302372/work
kaggle==1.7.4.2
kagglehub==0.3.10
keras==3.9.2
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
libclang==18.1.1
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1714510563922/work
mistune==3.0.2
mkl-fft @ file:///work/perseverance-python-buildout/croot/mkl_fft_1698845673361/work
mkl-random @ file:///work/perseverance-python-buildout/croot/mkl_random_1698845720894/work
mkl-service==2.4.0
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
narwhals==1.38.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.1
notebook_shim==0.2.4
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optax==0.2.4
optree==0.11.0
orbax-checkpoint==0.11.10
overrides==7.7.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==4.25.7
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.9.1
Pygments==2.18.0
pyHepMC3==3.2.6
pyHepMC3.search==3.2.6
pyparsing==3.1.2
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
redis @ file:///work/perseverance-python-buildout/croot/redis-py_1698900719407/work
referencing==0.35.1
requests @ file:///croot/requests_1707355572290/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==69.5.1
simple-parsing==0.1.7
simplejson @ file:///work/perseverance-python-buildout/croot/simplejson_1698901645930/work
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
streamlit==1.45.0
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.8
tensorflow-hub==0.16.1
tensorflow-metadata==1.17.1
tensorflow_decision_forests==1.12.0
tensorflowjs==4.22.0
tensorstore==0.1.72
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.19.0
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
toolz==1.0.0
torch==2.2.2
torchaudio==2.6.0
torchvision==0.17.2
tornado==6.4.1
tqdm @ file:///croot/tqdm_1714567712644/work
traitlets==5.14.3
treescope==0.1.9
triton==3.2.0
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
types-python-dateutil==2.9.0.20240316
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1715635851070/work
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockify @ file:///home/conda/feedstock_root/build_artifacts/websockify_1696158738258/work
Werkzeug==3.1.3
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt @ file:///work/perseverance-python-buildout/croot/wrapt_1698851040578/work
wurlitzer==3.1.1
ydf==0.11.0
zipp==3.21.0
zstandard @ file:///croot/zstandard_1714677652653/work
